+++ StackoverflowChatbot/Discord.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Text.RegularExpressions;
+using System.Linq;
 using System.Threading.Tasks;
 using Discord;
 using Discord.WebSocket;
@@ -92,6 +92,64 @@ namespace StackoverflowChatbot
 					}
 				}
 			}
+		}
+	
+		internal static IEnumerable<SocketGuildUser> GetUserByName(string name, bool onlyExactMatch = false)
+		{
+			var discordClient = Discord.GetDiscord().GetAwaiter().GetResult();
+			var guilds = discordClient.Guilds;
+			foreach (var guild in guilds)
+			{
+				guild?.DownloadUsersAsync();
+			}
+			var guildUsers = guilds.Select(guild => guild?.Users);
+			var result = new List<SocketGuildUser>();
+
+			foreach (var guild in guildUsers)
+			{
+				foreach (var user in guild)
+				{
+					var preferredName = user.Nickname;
+					if (user.Nickname == null)
+						preferredName = user.Username;
+					if(name.Equals(preferredName) || (!onlyExactMatch && (name.Contains(preferredName) || preferredName.Contains(name))))
+					{
+						result.Add(user);
+					}
+				}
+			}
+			
+			return result;
+		}
+
+		internal static IEnumerable<SocketRole> GetRolesByName(string name, bool onlyExactMatch = false)
+		{
+			var discordClient = Discord.GetDiscord().GetAwaiter().GetResult();
+			var rolesPerGuild = discordClient.Guilds.Select(guild => guild.Roles.Where(role => !onlyExactMatch ?
+				name.Contains(role.Name) || role.Name.Contains(name) : name.Equals(role.Name)));
+			
+			var result = new List<SocketRole>();
+			foreach (var roles in rolesPerGuild)
+			{
+				result.AddRange(roles);
+			}
+
+			return result;
+		}
+
+		internal static IEnumerable<SocketChannel> GetChannelsByName(string name, bool onlyExactMatch = false)
+		{
+			var discordClient = Discord.GetDiscord().GetAwaiter().GetResult();
+			var channelsPerGuild = discordClient.Guilds.Select(guild => guild.Channels.Where(channel => !onlyExactMatch ?
+				name.Contains(channel.Name) || channel.Name.Contains(name) : name.Equals(channel.Name)));
+
+			var result = new List<SocketChannel>();
+			foreach (var channels in channelsPerGuild)
+			{
+				result.AddRange(channels);
+			}
+
+			return result;
 		}
 	}
 }
